feat: update data types and improve error handling in collection and entry management

+++ microservice/gateway/handler/collection/collection.go
@@ -15,6 +15,6 @@ func New(i dao.Interface) *Api {
 }
 
 type CreateRequest struct {
-	TargetID   uint64 `json:"target_id" binding:"required"`
+	TargetID   uint32 `json:"target_id" binding:"required"`
 	TargetType uint32 `json:"target_type" binding:"required"`
 }
